backend (supabase)/requirements.txt
fastapi==0.110.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysql == 0.0.3
mysql-connector-python==8.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
uvicorn==0.29.0
Werkzeug==3.0.1
aiohttp>=3.7.4
boto3==1.34.72
